# Conflicts:
#	Extinguish/Extinguish/Countdown.cpp
#	Extinguish/Resources/Stats/Claptrap.stats
#	Extinguish/Resources/Stats/Slackbot.stats
#	Extinguish/Resources/Stats/Tom.stats
#	Extinguish/Resources/Stats/Wall-E.stats

+++ Extinguish/Extinguish/Countdown.cpp
@@ -79,7 +79,7 @@ void Countdown::Update(float dt)
 //misc//
 void Countdown::Reset()
 {
-	timer = timeTilPlay;
+	timer = (float)timeTilPlay;
 	canPlay = false;
 	curSecond = -1;
 	playSound = false;
